Working MAEbot library with AprilC library included

+++ src/lcmtypes/dynamixel_command_list_t.hpp
@@ -64,13 +64,13 @@ class dynamixel_command_list_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int dynamixel_command_list_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = (int64_t)getHash();
+    int64_t hash = getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -153,7 +153,7 @@ int dynamixel_command_list_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-uint64_t dynamixel_command_list_t::_computeHash(const __lcm_hash_ptr *p)
+int64_t dynamixel_command_list_t::_computeHash(const __lcm_hash_ptr *p)
 {
     const __lcm_hash_ptr *fp;
     for(fp = p; fp != NULL; fp = fp->parent)
@@ -161,7 +161,7 @@ uint64_t dynamixel_command_list_t::_computeHash(const __lcm_hash_ptr *p)
             return 0;
     const __lcm_hash_ptr cp = { p, (void*)dynamixel_command_list_t::getHash };
 
-    uint64_t hash = 0x131b5b70389b084fLL +
+    int64_t hash = 0x131b5b70389b084fLL +
          dynamixel_command_t::_computeHash(&cp);
 
     return (hash<<1) + ((hash>>63)&1);
