+++ src/lcmtypes/dynamixel_status_t.hpp
@@ -78,13 +78,13 @@ class dynamixel_status_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int dynamixel_status_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = (int64_t)getHash();
+    int64_t hash = getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -195,9 +195,9 @@ int dynamixel_status_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-uint64_t dynamixel_status_t::_computeHash(const __lcm_hash_ptr *)
+int64_t dynamixel_status_t::_computeHash(const __lcm_hash_ptr *)
 {
-    uint64_t hash = 0x73954bff708517ffLL;
+    int64_t hash = 0x73954bff708517ffLL;
     return (hash<<1) + ((hash>>63)&1);
 }
